+++ src/autobahn/twisted/rawsocket.py
@@ -295,7 +295,12 @@ class WampRawSocketProtocol(Int32StringReceiver):
         """
         Implements :func:`autobahn.wamp.interfaces.ITransport.abort`
         """
-        if self.isOpen():
+        # Guard on the transport, not isOpen(): isOpen() requires an attached
+        # WAMP session, but abort() must also tear down the transport when
+        # called before the session is established - e.g. on a failed opening
+        # handshake (bad magic byte / no suitable serializer), see #1850.
+        # Only a genuinely missing transport is a lost transport.
+        if self.transport is not None:
             if hasattr(self.transport, "abortConnection"):
                 # ProcessProtocol lacks abortConnection()
                 self.transport.abortConnection()
@@ -338,6 +343,7 @@ class WampRawSocketServerProtocol(WampRawSocketProtocol):
                         magic=_magic,
                     )
                     self.abort()
+                    return
                 else:
                     self.log.debug(
                         "WampRawSocketServerProtocol: correct magic byte received"
@@ -367,6 +373,7 @@ class WampRawSocketServerProtocol(WampRawSocketProtocol):
                         serializers=self.factory._serializers.keys(),
                     )
                     self.abort()
+                    return
 
                 # we request the client to send message of maximum length 2**reply_max_len_exp
                 #
@@ -460,6 +467,7 @@ class WampRawSocketClientProtocol(WampRawSocketProtocol):
                         magic=_LazyHexFormatter(self._handshake_bytes[0]),
                     )
                     self.abort()
+                    return
 
                 # peer requests us to _send_ messages of maximum length 2**max_len_exp
                 #
@@ -479,6 +487,7 @@ class WampRawSocketClientProtocol(WampRawSocketProtocol):
                         serializers=self._serializer.RAWSOCKET_SERIALIZER_ID,
                     )
                     self.abort()
+                    return
 
                 self._handshake_complete = True